Item photo upload via Library functional
Functionality to add photos to an item using the camera or device library is functional now after solving crashing bugs. Default attributes (value and description) now render to the screen when an item is selected, though UI updates will be needed. Delete item functionality fixed as photos were not being deleted from the database when deleted by the user.